+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Added
  - Warning when shadowing variable (#1366)
+ - Added support for compiler warnings (#1779)
 
 # v 2020.1 (2020-02-19) Changes in this release: